Joshua/texture compression work (#132)
* stuff

* stuff2

* Fixed the rotation

* Cleaned up warnings, added Rotate function to Transform, stubbed out RotateAround.

* Literally anything

* Further sprite work

* Animated sprites mostly working

* Fixed large level startup lag and SpriteText generally finished.

* Initial work on material working on all graphics types.

* Added basic easy_profile support

* Initial work on our own profiling macros.

* Changed some defines, profiling events.

* Switched to YTE macros, made some fixes to our defines

* Tiny bugfix for multiple animating sprites

* Minor platform cleanup

* Working on cmake

* Further cmake work

* Minor cmake fix

* Worked on texture compression

* Disable compression

* Remove crunch

* Brought in crunch and started working on better texture compression options.

* Optional texture compression

For some reason the progress callback isn't being called as often as I feel it should. Need to look into that.

* Disable model texture compression

* Bugfix for new Sprites.

+++ Source/YTE/Core/TestComponent.cpp
@@ -63,7 +63,7 @@ namespace YTE
     //
     //glm::vec3 scale{ mCurrentScale, mCurrentScale, mCurrentScale };
 
-    mCurrentRotation.y = mTime;
+    mCurrentRotation.y = static_cast<float>(mTime);
 
     //printf("%s: %f\n", mOwner->GetName().c_str(), mCurrentRotation.y);
